requirements.txt
streamlit==1.47.0
streamlit-folium
folium
requests
pandas
python-dotenv